+++ src/balances/balances.module.ts
@@ -2,12 +2,12 @@ import { Module } from '@nestjs/common';
 import { BalancesService } from './balances.service';
 import { BalancesController } from './balances.controller';
 import { TransactionApiModule } from '../datasources/transaction-api/transaction-api.module';
-import { ConfigServiceModule } from '../datasources/config-service/config-service.module';
+import { ConfigApiModule } from '../datasources/config-api/config-api.module';
 import { ExchangeModule } from '../datasources/exchange-api/exchange.module';
 
 @Module({
   controllers: [BalancesController],
   providers: [BalancesService],
-  imports: [TransactionApiModule, ExchangeModule, ConfigServiceModule],
+  imports: [TransactionApiModule, ExchangeModule, ConfigApiModule],
 })
 export class BalancesModule {}
