+++ src/balances/balances.service.ts
@@ -4,9 +4,9 @@ import { ExchangeApi } from '../datasources/exchange-api/exchange.service';
 import { TokenInfo } from '../common/entities/tokeninfo.entity';
 import { TokenType } from '../common/entities/tokentype.entity';
 import { Balances } from './entities/balances.entity';
-import { ConfigService } from '../datasources/config-service/config-service.service';
+import { ConfigApi } from '../datasources/config-api/config-api.service';
 import { TransactionApiManager } from '../datasources/transaction-api/transaction-api.manager';
-import { NativeCurrency } from '../datasources/config-service/entities/native.currency.entity';
+import { NativeCurrency } from '../datasources/config-api/entities/native.currency.entity';
 import { Balance } from './entities/balance.entity';
 
 @Injectable()
@@ -14,7 +14,7 @@ export class BalancesService {
   static readonly fromRateCurrencyCode: string = 'USD';
 
   constructor(
-    private readonly safeConfigService: ConfigService,
+    private readonly configApi: ConfigApi,
     private readonly safeTransactionManager: TransactionApiManager,
     private readonly exchangeApi: ExchangeApi,
   ) {}
@@ -34,7 +34,7 @@ export class BalancesService {
       BalancesService.fromRateCurrencyCode,
     );
     const nativeCurrency: NativeCurrency = (
-      await this.safeConfigService.getChain(chainId)
+      await this.configApi.getChain(chainId)
     ).nativeCurrency;
 
     // Map balances payload
