+++ src/chains/chains.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ChainsService } from './chains.service';
 import { ChainsController } from './chains.controller';
-import { ConfigServiceModule } from '../datasources/config-service/config-service.module';
+import { ConfigApiModule } from '../datasources/config-api/config-api.module';
 import { TransactionApiModule } from '../datasources/transaction-api/transaction-api.module';
 
 @Module({
-  imports: [ConfigServiceModule, TransactionApiModule],
+  imports: [ConfigApiModule, TransactionApiModule],
   controllers: [ChainsController],
   providers: [ChainsService],
 })
