+++ src/datasources/config-api/entities/native.currency.entity.ts
